move utils to utils_mod
