+++ manager/utils_mod/directory_class.py
